Reported compile/segfault fixes
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@337 37977431-3df6-0310-b722-df95706aa16b

+++ rts/BFGroundDrawer.cpp
@@ -960,7 +960,7 @@ bool CBFGroundDrawer::UpdateTextures()
 						m=md->moveMath->SpeedMod(*md, x*2,y*2);
 						if(gs->cheatEnabled && md->moveMath->IsBlocked2(*md, x*2+1, y*2+1) & (CMoveMath::BLOCK_STRUCTURE | CMoveMath::BLOCK_TERRAIN))
 							m=0;
-						m=min(1.,sqrt(m));
+						m=min(1.,(double)sqrt(m));
 					}
 					infoTexMem[a*4+0]=255-int(m)*255;
 					infoTexMem[a*4+1]=int(m)*255;
